문정) 카테고리 수정

+++ src/main/java/com/example/mall/controller/CategoryController.java
@@ -8,6 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.example.mall.service.CategoryService;
 import com.example.mall.util.TeamColor;
@@ -55,7 +56,18 @@ public class CategoryController { // Author : 김동현
 	
 	// removeCategory
 	@GetMapping("/staff/removeCategory")
-	public String removeCategory(@RequestParam Integer categoryNo) {
+	public String removeCategory(@RequestParam Integer categoryNo, RedirectAttributes redirectAttributes, Model model) {
+		
+		// 해당 카테고리에 상품이 존재하면 삭제 불가
+		Integer result = categoryService.selectCategoryGoods(categoryNo);
+		
+		if(result != 0) {
+			redirectAttributes.addFlashAttribute("categoryMsg", "해당 카테고리에 상품이 존재하여 삭제 불가합니다.");
+			
+			
+			log.debug(TeamColor.KMJ + "해당 카테고리에 상품이 존재하여 삭제 불가합니다." + TeamColor.RESET);
+			return "redirect:/staff/getCategoryList";
+		}
 		
 		log.debug(TeamColor.KDH + "removeCategoryNo : " + categoryNo + TeamColor.RESET); // debug
 		int removeCategoryRow = categoryService.removeCategory(categoryNo);
